Allow installation on devices without GPS/network location
The Google Play Store interprets the permissions ACCESS_FINE_LOCATION and ACCESS_COARSE_LOCATION as implying that GPS and network location are required, respectively.  With the External I/O plugin installed, though, Avare is useful on devices without GPS, so we should specify that these hardware features are not required.