Closes #372
This is a requested user feature.

Addresses the issue that ICAO codes were called decimal, although they are hexidecimal.  The summary in the preferences field is confusing.

Addresses the issue that user input is not capitalized in the preferences and FAA form. The form is already capitalized on submission, but this is not apparent to the user.

An additional class very similar to an existing class was added to match the existing Android interface.  Note: this interface is deprecated.