Update help.html
Thanks to lots of excellent work by Sam Hoskins, I was inspired to potentially mess that work up by editing it with my own "improvements" since he's on vacation. Presumably before this replaces the in-app Help, he and others will review it further. Even if it's added as-is after ensuring it doesn't somehow mess up the app, I'm confident it's a major improvement on the inconsistent and error-ridden file that has evolved in the years since a last full review. I'm burnt out, and probably won't find time to edit again in the near term. So feel free to munch away at it yourself. I'm going to try attaching Sam's Help-2 file that I've edited.